pyproject.toml
[project]
name = "vaf-am-ingestion"
version = "0.1.0"
description = "VAF AM Build 01 — Multi-Source Data Ingestion"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.28.0",
    "feedparser>=6.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "beautifulsoup4>=4.12.0",
    "aiosqlite>=0.19.0",
    "tiktoken>=0.7.0",
    "pdfplumber>=0.10.0",
]
